% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{calc_crosstables}
\alias{calc_crosstables}
\title{Calculate the CrossTable for the given variables.}
\usage{
calc_crosstables(data, ...)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{...}{Sets of variables to analyse (see above).}
}
\value{
list
}
\description{
Does not do much more than CrossTable does but saves some keystrokes when
doing a lot of CrossTables.

It takes sets of the form c(ID, X, Y, SHOW), where SHOW specifies whether the
crosstable should be shown.
The output of \code{gmodels::CrossTable(data[[X]], data[[Y]], fisher = TRUE, expected = TRUE)} will be saved in the returned list \code{list[[ID]]}.
}
