% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{calc_cv}
\alias{calc_cv}
\title{Calculate the coefficient of variation of \code{x}.}
\usage{
calc_cv(x)
}
\arguments{
\item{x}{The values to use.}
}
\value{
The cv.
}
\description{
Calculate the coefficient (cv) of variation of the values in vector \code{x}.
}
