% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{calc_sequential_difference}
\alias{calc_sequential_difference}
\title{Calculate the difference between sequential rows while grouping.}
\usage{
calc_sequential_difference(data, values, arrange, ...,
  signif_above = 0.2)
}
\arguments{
\item{data}{A tibble.}

\item{values}{The column with the values to calculate the difference with.}

\item{arrange}{The column to arrange the data by.}

\item{...}{The columns to group by.}

\item{signif_above}{An difference above this cut off is considered significant.}
}
\value{
A tibble.
}
\description{
Calculate the difference between sequential rows while grouping.
}
