% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristics.R
\name{characteristic_get}
\alias{characteristic_get}
\title{Get metrics for a given characteristic of your dataset.}
\usage{
characteristic_get(data, characteristic, ..., name = NULL,
  events = c(TRUE), quantiles = c("q5", "q25", "q50", "q75", "q95"),
  force_count = FALSE, template = NULL, decimals = eenv_decimals,
  decimals_perc = FALSE)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{characteristic}{The column containing the characteristic to analyse.}

\item{...}{Subgroups for which to do a separate calculation.}

\item{name}{The first item in vector (only if a \code{template} string is given).}

\item{events}{The event(s) to count (only if characteristic is not numeric or \code{force_count}).}

\item{quantiles}{The quantiles to collect (default: \code{c("q5", "q25", "q50", "q75", "q95")},
only if characteristic is numeric).}

\item{force_count}{Count frequency of characteristic although its numeric?}

\item{template}{Apply a template string.}

\item{decimals}{The number of decimals to display.}

\item{decimals_perc}{Display decimals for percentages.}
}
\value{
A named vector.
}
\description{
This function gives similar results to \code{summary()}.
}
