% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristics.R
\name{characteristics_table}
\alias{characteristics_table}
\title{Format several characteristics as a table.}
\usage{
characteristics_table(..., headers = "NAME")
}
\arguments{
\item{...}{Vectors of the same length to be coerced to a table.}

\item{headers}{The column in which to insert subsection-headers}
}
\value{
A tibble
}
\description{
Format several characteristics as a table.
}
