% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{colour_base}
\alias{colour_base}
\title{Default colour for dots / axes / text ... (dark grey).}
\format{An object of class \code{character} of length 1.}
\usage{
colour_base
}
\description{
Default colour for dots / axes / text ... (dark grey).
}
\seealso{
Other theme-related functions and objects: \code{\link{colour_palette}},
  \code{\link{colour_secondary}},
  \code{\link{colour_tert}}, \code{\link{create_theme}},
  \code{\link{eenv_theme}}

Other variables and functions for plotting: \code{\link{colour_bg}},
  \code{\link{colour_palette}},
  \code{\link{colour_secondary}},
  \code{\link{colour_tert}}, \code{\link{eenv_plot_dim}},
  \code{\link{save_plots}}
}
\keyword{datasets}
