% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{colour_bg}
\alias{colour_bg}
\title{Default background colour.}
\format{An object of class \code{character} of length 1.}
\usage{
colour_bg
}
\description{
Default background colour.
}
\seealso{
Other variables and functions for plotting: \code{\link{colour_base}},
  \code{\link{colour_palette}},
  \code{\link{colour_secondary}},
  \code{\link{colour_tert}}, \code{\link{eenv_plot_dim}},
  \code{\link{save_plots}}
}
\concept{theme-related functions and objects (white)}
\concept{variables and functions for plotting}
\keyword{datasets}
