% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{colour_secondary}
\alias{colour_secondary}
\title{The colour to use if more than 1 colours are needed (brighter grey).}
\format{An object of class \code{character} of length 1.}
\usage{
colour_secondary
}
\description{
The colour to use if more than 1 colours are needed (brighter grey).
}
\seealso{
Other theme-related functions and objects: \code{\link{colour_base}},
  \code{\link{colour_palette}}, \code{\link{colour_tert}},
  \code{\link{create_eenv_theme}}, \code{\link{eenv_theme}}

Other variables and functions for plotting: \code{\link{colour_base}},
  \code{\link{colour_bg}}, \code{\link{colour_palette}},
  \code{\link{colour_tert}}, \code{\link{eenv_plot_dim}},
  \code{\link{save_plots}}
}
\keyword{datasets}
