% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{colour_tert}
\alias{colour_tert}
\title{The colour to use if more than 2 colours are needed (even brighter grey).}
\format{An object of class \code{character} of length 1.}
\usage{
colour_tert
}
\description{
The colour to use if more than 2 colours are needed (even brighter grey).
}
\seealso{
Other theme-related functions and objects: \code{\link{colour_base}},
  \code{\link{colour_palette}},
  \code{\link{colour_secondary}},
  \code{\link{create_eenv_theme}}, \code{\link{eenv_theme}}

Other variables and functions for plotting: \code{\link{colour_base}},
  \code{\link{colour_bg}}, \code{\link{colour_palette}},
  \code{\link{colour_secondary}},
  \code{\link{eenv_plot_dim}}, \code{\link{save_plots}}
}
\keyword{datasets}
