% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{create_eenv_theme}
\alias{create_eenv_theme}
\title{Create a conservative theme for ggplot2.}
\usage{
create_eenv_theme(base_size = 10, base_family = "",
  base_colour_fg = colour_base, base_colour_bg = colour_bg,
  base_colour_sec = colour_tert, legend_pos = "bottom")
}
\arguments{
\item{base_size}{The font size in points used for the text.}

\item{base_family}{The font-family.}

\item{base_colour_fg}{The colour used for lines / dots / axes.}

\item{base_colour_bg}{The background colour.}

\item{base_colour_sec}{The colour used for the strips in facet_wraps, ... .}

\item{legend_pos}{Where to put the legend ("top", "right", "left", "bottom").}
}
\value{
The theme.
}
\description{
Create a conservative theme for ggplot2.
}
\seealso{
Other theme-related functions and objects: \code{\link{colour_base}},
  \code{\link{colour_palette}},
  \code{\link{colour_secondary}},
  \code{\link{colour_tert}}, \code{\link{eenv_theme}}
}
