% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eenv.R
\docType{package}
\name{eenv-package}
\alias{eenv}
\alias{eenv-package}
\title{A package to load packages, functions and variables I frequently use.}
\description{
Loads packages, adds global variables, themes and functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/randomchars42/eenv}
  \item Report bugs at \url{https://github.com/randomchars42/eenv/issues}
}

}
\author{
\strong{Maintainer}: Eike Christian Kühn \email{eike.kuehn@pixelwoelkchen.de}

}
