% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{eenv_decimals}
\alias{eenv_decimals}
\title{The number of digits to round to (not for p-values).}
\format{An object of class \code{numeric} of length 1.}
\usage{
eenv_decimals
}
\description{
The number of digits to round to (not for p-values).
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{eenv_signif_bottom}},
  \code{\link{eenv_signif_digits}},
  \code{\link{format_float}}, \code{\link{format_int}},
  \code{\link{format_number}}, \code{\link{format_pad}},
  \code{\link{format_perc}}, \code{\link{format_p}}
}
\concept{variables and functions for displaying numbers}
\keyword{datasets}
