% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{eenv_plot_dim}
\alias{eenv_plot_dim}
\title{Default settings for \code{\link[=save_plots]{save_plots()}}.}
\format{An object of class \code{list} of length 8.}
\usage{
eenv_plot_dim
}
\description{
List:
\itemize{
\item dpi -> The resolution to use in DPI (dots per inch).
\item format -> Formats to save plots in (plots can be saved in more than format
simultaneously).
\item units -> Units for plot-size.
\item width -> Different steps of widht the plot may have.
\item height -> Height of one single plot.
\item columns_max -> When plotting multiple plots onto one grid, the number of
columns to use
\item axes -> Which axes to align (see \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}]).
\item align -> Which directions to use for alignment.
}
}
\seealso{
\code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}

Other variables and functions for plotting: \code{\link{colour_base}},
  \code{\link{colour_bg}}, \code{\link{colour_palette}},
  \code{\link{colour_secondary}},
  \code{\link{colour_tert}}, \code{\link{save_plots}}
}
\concept{variables and functions for plotting}
\keyword{datasets}
