% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{eenv_signif_digits}
\alias{eenv_signif_digits}
\title{The number of non-zero digits to display for p-values.}
\format{An object of class \code{numeric} of length 1.}
\usage{
eenv_signif_digits
}
\description{
The number of non-zero digits to display for p-values.
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{eenv_decimals}},
  \code{\link{eenv_signif_bottom}},
  \code{\link{format_float}}, \code{\link{format_int}},
  \code{\link{format_number}}, \code{\link{format_pad}},
  \code{\link{format_perc}}, \code{\link{format_p}}
}
\keyword{datasets}
