% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{eenv_signif_steps}
\alias{eenv_signif_steps}
\title{Steps of p-values used for the number of stars in plots or tables.}
\format{An object of class \code{numeric} of length 4.}
\usage{
eenv_signif_steps
}
\description{
Steps of p-values used for the number of stars in plots or tables.
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{eenv_decimals}},
  \code{\link{eenv_signif_bottom}},
  \code{\link{eenv_signif_digits}},
  \code{\link{format_float}}, \code{\link{format_int}},
  \code{\link{format_number}}, \code{\link{format_pad}},
  \code{\link{format_perc}}, \code{\link{format_p}}
}
\concept{variables and functions for displaying numbers}
\keyword{datasets}
