% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{eenv_theme}
\alias{eenv_theme}
\title{A conservative theme with default colours in black and white.}
\format{An object of class \code{list} of length 3.}
\usage{
eenv_theme
}
\description{
A conservative theme with default colours in black and white.
}
\seealso{
Other theme-related functions and objects: \code{\link{colour_base}},
  \code{\link{colour_palette}},
  \code{\link{colour_secondary}},
  \code{\link{colour_tert}}, \code{\link{create_theme}}
}
\keyword{datasets}
