% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure}
\alias{exposure}
\alias{exposure_get_relation}
\title{Work with exposure vs. conditions.}
\usage{
exposure_get_relation(data, exposure, condition, exposure_targ = TRUE,
  condition_targ = TRUE, alpha = eenv_alpha,
  haldane_anscombe_correction = TRUE, relations = c("OR", "RR", "ARR"),
  confusion_matrix = NULL)
}
\arguments{
\item{data}{A tibble holding the data}

\item{exposure}{The column holding the predicted conditions / exposure.}

\item{condition}{The column holding the observed conditions.}

\item{exposure_targ}{The value that signifies the exposed condition.}

\item{condition_targ}{The value that signifies the targeted observed
condition.}

\item{alpha}{The alpha level.}

\item{haldane_anscombe_correction}{Apply Haldane Anscombe correction if}

\item{relations}{Vector with at least one in \code{c("OR", "RR", "ARR")}.
necessary.}

\item{confusion_matrix}{A confusion matrix as returned by
\code{test_get_confusion_matrix}, \code{data} will be ignored.}
}
\value{
variable
}
\description{
Functions to retrieve risk ratio, difference in absolute risk and odds ratio.

Structure: Suppose you want to compare a group of people exposed to x vs. a
group without exposure:

|   | exposure    | observed    | comment           |
|---|-------------|-------------|-------------------|
| 1 | pos         | yes         | <- true positive  |
| 2 | pos         | no          | <- false positive |
| 3 | neg         | no          | <- true negative  |
| 4 | pos         | yes         | ...               |
| 5 | neg         | no          | ...               |
}
