% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure_format_relations}
\alias{exposure_format_relations}
\title{Format the output of \code{exposure_get_relation}.}
\usage{
exposure_format_relations(raw_relations, relations = c("OR", "RR",
  "ARR"))
}
\arguments{
\item{raw_relations}{The output of \code{test_get_relation}.}

\item{relations}{Vector with at least one in \code{c("OR", "RR", "ARR")}.}
}
\value{
named vector
}
\description{
Format the output of \code{exposure_get_relation}.
}
