% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{extract_p_values_from_result_list}
\alias{extract_p_values_from_result_list}
\title{Extract p-values from a result list.}
\usage{
extract_p_values_from_result_list(result_list, names_num = FALSE)
}
\arguments{
\item{result_list}{The result list.}

\item{names_num}{Convert the group names to numeric.}
}
\value{
vector
}
\description{
Convenience function for extracting p-values from result lists as returned by
\code{test_wilcoxon_for_groups} or \code{test_friedman}. The result is vector with the
group names as names and can be used by \code{ggplot_annotate_signif}.
}
