% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_float}
\alias{format_float}
\title{Round and format a floating point number.}
\usage{
format_float(x, decimals = eenv_decimals)
}
\arguments{
\item{x}{The number to format.}

\item{decimals}{The number of decimals to round to (not p-values).}
}
\value{
A string.
}
\description{
Round and format a floating point number.
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{eenv_decimals}},
  \code{\link{eenv_signif_bottom}},
  \code{\link{eenv_signif_digits}},
  \code{\link{eenv_signif_steps}},
  \code{\link{format_int}}, \code{\link{format_number}},
  \code{\link{format_pad}}, \code{\link{format_perc}},
  \code{\link{format_p}}
}
\concept{variables and functions for displaying numbers}
