% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_int}
\alias{format_int}
\title{Format an integer for displaying.}
\usage{
format_int(x)
}
\arguments{
\item{x}{The number to format.}
}
\value{
A string.
}
\description{
Format an integer for displaying.
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{format_float}},
  \code{\link{format_number}}, \code{\link{format_pad}},
  \code{\link{format_perc}}, \code{\link{format_p}},
  \code{\link{global_decimals}},
  \code{\link{global_signif_bottom}},
  \code{\link{global_signif_digits}}
}
