% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_number}
\alias{format_number}
\title{Format a number for displaying.}
\usage{
format_number(x, type = "float", decimals = global_decimals,
  signif_digits = global_signif_digits,
  signif_bottom = global_signif_bottom, force_perc_decimals = FALSE)
}
\arguments{
\item{x}{The number to format.}

\item{type}{The type of number ("int", "perc", "float", "p").}

\item{decimals}{The number of decimals to round to (not p-values).}

\item{signif_digits}{The number of non-zero digits to display for p-values.}

\item{signif_bottom}{The number below which "< NUMBER" is displayed for
p-values.}

\item{force_perc_decimals}{TRUE if decimals should be displayed for
percentages.}
}
\value{
A string.
}
\description{
This function respects \link{global_signif_digits}, \link{global_signif_bottom},
\link{global_decimals} as default.
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{format_float}},
  \code{\link{format_int}}, \code{\link{format_pad}},
  \code{\link{format_perc}}, \code{\link{format_p}},
  \code{\link{global_decimals}},
  \code{\link{global_signif_bottom}},
  \code{\link{global_signif_digits}}
}
