% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_p}
\alias{format_p}
\title{Format a p-value.}
\usage{
format_p(x, signif_digits = global_signif_digits,
  signif_bottom = global_signif_bottom)
}
\arguments{
\item{x}{The number to format.}

\item{signif_digits}{The number of non-zero digits to display for p-values.}

\item{signif_bottom}{The number below which "< NUMBER" is displayed for
p-values.}
}
\value{
A string.
}
\description{
Format a p-value.
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{format_float}},
  \code{\link{format_int}}, \code{\link{format_number}},
  \code{\link{format_pad}}, \code{\link{format_perc}},
  \code{\link{global_decimals}},
  \code{\link{global_signif_bottom}},
  \code{\link{global_signif_digits}}
}
