% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_pad}
\alias{format_pad}
\title{Pad a string to a given length.}
\usage{
format_pad(x, width, side = "left", character = " ")
}
\arguments{
\item{x}{The string to pad.}

\item{width}{How long should the string be?}

\item{side}{To which side should the padding be added ("left" / "right")?}

\item{character}{Which character to use.}
}
\value{
A string.
}
\description{
Pad a string to a given length.
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{format_float}},
  \code{\link{format_int}}, \code{\link{format_number}},
  \code{\link{format_perc}}, \code{\link{format_p}},
  \code{\link{global_decimals}},
  \code{\link{global_signif_bottom}},
  \code{\link{global_signif_digits}}
}
