% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_perc}
\alias{format_perc}
\title{Format a percentage for displaying.}
\usage{
format_perc(x, decimals = eenv_decimals, force_perc_decimals = FALSE)
}
\arguments{
\item{x}{The number to format.}

\item{decimals}{The number of decimals to round to (not p-values).}

\item{force_perc_decimals}{TRUE if decimals should be displayed for
percentages.}
}
\value{
A string.
}
\description{
Format a percentage for displaying.
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{eenv_decimals}},
  \code{\link{eenv_signif_bottom}},
  \code{\link{eenv_signif_digits}},
  \code{\link{format_float}}, \code{\link{format_int}},
  \code{\link{format_number}}, \code{\link{format_pad}},
  \code{\link{format_p}}
}
