% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_variable_by_groups}
\alias{get_variable_by_groups}
\title{Turn long data with subgroups into wide data.}
\usage{
get_variable_by_groups(data, column, group, id, ...)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{column}{Which column to use.}

\item{group}{The column that parts the variable into subgroups.}

\item{id}{A column holding ids to identify the same entity in each subgroup.}

\item{...}{Further ids.}
}
\value{
list
}
\description{
Turns data in a column "column" which is divided into groups by a column
"group" into a tibble with a column for each group (each unique value in
"group").
}
