% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{ggplot_annotate_signif}
\alias{ggplot_annotate_signif}
\title{Extract p-values from a result list.}
\usage{
ggplot_annotate_signif(data, x, y, p_values, label = "stars",
  margin = 1, static_y = TRUE, steps = c(0.05, 0.01, 0.001),
  font_face = "plain", font_size = eenv_theme[[1]]$text$size)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{x}{The column used for the x-axis.}

\item{y}{The column used for the y-axis.}

\item{p_values}{A vector containing p-values, the name marks the
corresponding x.}

\item{label}{Either "stars" or "p-values".}

\item{margin}{The distance above the highest y at x in units of y.}

\item{static_y}{Use the maximum y + margin for all x.}

\item{steps}{If \code{label} is "stars" add an "*" for each step teh p_value is
below.}

\item{font_face}{The font face used for the labels ("plain"|"bold"|"italic"|
"bold.italic")}

\item{font_size}{The font size in points.}
}
\value{
list
}
\description{
Convenience function for extracting p-values from result lists as returned by
\code{test_wilcoxon_for_groups} or \code{test_friedman}. The result is vector with the
group names as names and can be used by \code{ggplot_annotate_signif}.
}
