% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{global_alpha}
\alias{global_alpha}
\title{The default alpha used in functions of this package.}
\format{An object of class \code{numeric} of length 1.}
\usage{
global_alpha
}
\description{
The default alpha used in functions of this package.
}
\keyword{datasets}
