% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{global_confidence_level}
\alias{global_confidence_level}
\title{The default confidence level used in functions of this package.}
\format{An object of class \code{numeric} of length 1.}
\usage{
global_confidence_level
}
\description{
The default confidence level used in functions of this package.
}
\keyword{datasets}
