% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{global_decimals}
\alias{global_decimals}
\title{The number of digits to round to (not for p-values).}
\format{An object of class \code{numeric} of length 1.}
\usage{
global_decimals
}
\description{
The number of digits to round to (not for p-values).
}
\seealso{
Other variables and functions for displaying numbers: \code{\link{format_float}},
  \code{\link{format_int}}, \code{\link{format_number}},
  \code{\link{format_pad}}, \code{\link{format_perc}},
  \code{\link{format_p}},
  \code{\link{global_signif_bottom}},
  \code{\link{global_signif_digits}}
}
\keyword{datasets}
