% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_number}
\alias{is_number}
\title{Check if \code{x} is a number.}
\usage{
is_number(x)
}
\arguments{
\item{x}{The variable to check.}
}
\value{
\code{TRUE} / \code{FALSE}
}
\description{
Check that \code{x} is neither \code{NULL}, nor \code{NA`` and is of type }numeric\code{or}integer`.
}
