% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{package_available}
\alias{package_available}
\title{Test if a package is installed.}
\usage{
package_available(package)
}
\arguments{
\item{package}{The name of the package as a string.}
}
\value{
TRUE if the package is installed.
}
\description{
Test if a package is installed.
}
\seealso{
Other package functions: \code{\link{package_load}}
}
\concept{package functions}
