% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{package_load}
\alias{package_load}
\title{Load a package.}
\usage{
package_load(package)
}
\arguments{
\item{package}{The name of the package as a string.}
}
\value{
TRUE if the package could be (installed and) loaded.
}
\description{
If the package is not installed. Try to install it before loading.
}
\seealso{
Other package functions: \code{\link{package_available}}
}
\concept{package functions}
