% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosets.R
\name{plate_check_discrepant_multiples}
\alias{plate_check_discrepant_multiples}
\title{Search for and remove multiples which sets.}
\usage{
plate_check_discrepant_multiples(data, ids, ..., threshold = 0.2,
  throw_warnings = TRUE, remove = TRUE, set = set)
}
\arguments{
\item{data}{Tibble holding the data.}

\item{ids}{The column that holds the ids of the samples.}

\item{...}{Columns to use for filtering.}

\item{threshold}{The threshold used for filtering:
e.g. \code{concentration_cv} (coefficient of variation) > \code{0.2}. If a column
\code{recovery} is found the threshold for warning will be:
\code{recovery < 1 - threshold | recovery > 1 + threshold}}

\item{throw_warnings}{Throw a warning for each detected / removed sample?}

\item{remove}{Remove detected samples?}

\item{set}{The column that holds the id of the sets / plates.}
}
\value{
The tibble.
}
\description{
Dealing with raw data one might have to exclude a sample where duplicates /
triplicates / ... are too discrepant to use. This function identifies those
samples and might delete them.
}
