% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosets.R
\name{plates_read}
\alias{plates_read}
\title{Read sets (plates) and calculate concentrations and variability.}
\usage{
plates_read(plates, cal_names, cal_values, exclude_cals = list(),
  additional_vars = c("var_name"), additional_sep = "_", sep = ",",
  dec = "AUTO", path = ".", file_name = "plate_#NUM#.csv",
  model_func = bioset::fit_lnln, plot_func = bioset::plot_lnln,
  interpolate_func = bioset::interpolate_lnln, write_data = TRUE,
  use_written_data = FALSE)
}
\arguments{
\item{plates}{The number of plates (e.g. \code{3`` attempts to read }plate_1.csv\code{, }plate_2.csv\code{, }plate_3.csv\code{), see }file_name`.}

\item{cal_names}{A vector of strings containing the names of the samples used
as calibrators.}

\item{cal_values}{A numeric vector with the known concentrations of those
samples (must be in the same order).}

\item{exclude_cals}{A list of calibrators to exclude, e.g.:
\code{list(plate1 = c("CAL1"))}.}

\item{additional_vars}{Vector of strings containing the names for the
additional columns.}

\item{additional_sep}{String / RegExp that separates additional vars, e.g.:
\code{"ID_blue_cold"} with \code{additional_sep = "_"} will be separated
into three columns containing \code{"ID"}, \code{"blue"} and \code{"cold"}.
If the separated data would exceed the columns in \code{additional_vars}
the last column will contain a string with separator (e.g.: \code{"blue_cold"}).
If data is missing \code{NA} is inserted.}

\item{sep}{Separator used in the csv-file, either "," or ";" (see
\code{\link[utils:read.csv]{utils::read.csv()}}).}

\item{dec}{The character used for decimal points (see \code{\link[utils:read.csv]{utils::read.csv()}}).
"AUTO" will result in "." if \code{sep} is "," and "," for ";".}

\item{path}{The path to the file (no trailing "/" or "\\" !).}

\item{file_name}{Naming scheme for the files. The default is
\code{plate_#NUM#.csv}, where \code{#NUM#} gets replaced by the number of the plates,
see \code{plates}. The filename must contain \code{#NUM#}.}

\item{model_func}{A function generating a model to fit the calibrators,
e.g. \code{\link[=fit_linear]{fit_linear()}}, \code{\link[=fit_lnln]{fit_lnln()}}.}

\item{plot_func}{Function used to display the fitted line.}

\item{interpolate_func}{A function used to interpolate the concentrations of
the other samples, based on the model, e.g.
\code{\link[=interpolate_linear]{interpolate_linear()}}, \code{\link[=interpolate_lnln]{interpolate_lnln()}}.}

\item{write_data}{Write the calculated data into \code{data_all.csv} and
\code{data_samples.csv}?}

\item{use_written_data}{Try to read \code{data_all.csv} and \code{data_read.csv}
instead of raw data. Useful if you have to re-run the script, but the raw
data does not change.}
}
\value{
A list of params.
}
\description{
Writes the processed data into two files: \code{data_samples.csv},
\code{data_all.csv} and returns a list containing:
\itemize{
\item all: all rows including duplicate and calibrators.
\item samples: samples only, no calibrators, no duplicates.
\item plateNUMBER:
\itemize{
\item plot: plot of the calibrators
\item model: model used to fit a line to the calibrators
}
}
}
