% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Calculate and plot a qq-plot.}
\usage{
plot_qq(data, column, method = "middle")
}
\arguments{
\item{data}{A tibble containing the data.}

\item{column}{Which column to use.}

\item{method}{How to calculate ranks.}
}
\value{
list
}
\description{
Calculate and plot a qq-plot with actual values as axes instead of z-values.
}
