% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{plot_qq_for_groups}
\alias{plot_qq_for_groups}
\title{Calculate and plot a qq-plot for all groups.}
\usage{
plot_qq_for_groups(data, column, group, id, method = "middle")
}
\arguments{
\item{data}{A tibble containing the data.}

\item{column}{Which column to use.}

\item{group}{The column that parts the variable into subgroups.}

\item{id}{A column holding ids to identify the same entity in each subgroup.}

\item{method}{How to calculate ranks.}
}
\value{
list
}
\description{
Calculate and plot a qq-plot with actual values as axes instead of z-values
for all groups. Useful when checking series of measurements.
}
