% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{scan_crosstables-deprecated}
\alias{scan_crosstables-deprecated}
\title{Calculate all CrossTables for the given variables.}
\arguments{
\item{data}{A tibble containing the data.}

\item{variables_x}{A list of variable names as strings.}

\item{variables_y}{A list of variable names as strings.}
}
\value{
list
}
\description{
Give an overview of possible correlations by calculating all CrossTables for
the given variables.
}
\seealso{
\code{eenv_deprecated}
}
\keyword{internal}
