% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{scan_crosstables}
\alias{scan_crosstables}
\title{Calculate all CrossTables for the given variables.}
\usage{
scan_crosstables(data, variables_x, variables_y)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{variables_x}{A list of variable names as strings.}

\item{variables_y}{A list of variable names as strings.}
}
\value{
list
}
\description{
Give an overview of possible correlations by calculating all CrossTables for
the given variables.
}
