% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{scan_rocs}
\alias{scan_rocs}
\title{Calculate all Reciever-Operator-Curves (ROCs).}
\usage{
scan_rocs(data, predictors, responses, response_pos = TRUE)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{predictors}{A list of variable names.}

\item{responses}{A list of variable names.}

\item{response_pos}{The response that will be interpreted as an event.}
}
\description{
Calculate all Reciever-Operator-Curves (ROCs) and the corresponding area
under the curve (AUC).
}
