% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosets.R
\name{sets_read}
\alias{sets_read}
\title{Read sets (plates) and calculate concentrations and variability.}
\usage{
sets_read(plates, cal_names, cal_values, exclude_cals = c(),
  additional_vars = c("name"), sep = ",", path = ".",
  model_func = bioset::fit_lnln, plot_func = bioset::plot_lnln,
  interpolate_func = bioset::interpolate_lnln)
}
\arguments{
\item{plates}{The number of plates (e.g. 3 attempts to read "plate_1.csv",
"plate_2.csv", "plate_3.csv")}

\item{cal_names}{A vector of strings containing the names of the samples used
as calibrators,}

\item{cal_values}{A numeric vector with the known concentrations of those
samples (must be in the same order).}

\item{exclude_cals}{A list of calibrators to exclude, e.g.:
\code{list(plate1 = c("CAL1"))}.}

\item{additional_vars}{Vector of strings containing the names for the
additional columns.}

\item{sep}{Separator used in the csv-file, either "," or ";" (see
\code{\link[utils]{read.csv}})}

\item{path}{The path to the file (no trailing "/" or "\/"!).}

\item{model_func}{A function generating a model to fit the calibrators,
e.g. \code{\link{fit_linear}}, \code{\link{fit_lnln}}}

\item{plot_func}{Function used to display the fitted line.}

\item{interpolate_func}{A function used to interpolate the concentrations of
the other samples, based on the model, e.g.
\code{\link{interpolate_linear}}, \code{\link{interpolate_lnln}}}
}
\value{
A list of params.
}
\description{
Writes the processed data into two files: \code{data_samples.csv},
\code{data_all.csv} and returns a list containing:
\itemize{
\item all: all rows including duplicate and calibrators.
\item samples: samples only, no calibrators, no duplicates.
\item plateNUMBER:
\itemize{
\item plot: plot of the calibrators
\item model: model used to fit a line to the calibrators
}
}
}
