% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{test_friedman}
\alias{test_friedman}
\title{Friedman's test for a series of data.}
\usage{
test_friedman(data, column, group, id, reference_group,
  ignore_friedman = FALSE)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{column}{Which column to use.}

\item{group}{The column that parts the variable into subgroups.}

\item{id}{A column holding ids to identify the same entity in each subgroup.}

\item{reference_group}{The value in column \code{group} that indicates the
reference group.}

\item{ignore_friedman}{Ignore the result of Friedman's test and compare all
subgroups.}
}
\value{
list
}
\description{
Calculate Friedman's test for a series of data. If it shows that the
groups in the series are not equal it will calculate a Wilcoxon's signed-
rank test comparing each group to the reference group.
}
