% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{test_simple_predictions}
\alias{test_simple_predictions}
\title{Calculate a crosstable or ROC and test metrics for the given variables.}
\usage{
test_simple_predictions(data, ..., alpha = eenv_alpha, show = FALSE,
  force_show = c(), print_auc = FALSE, print_points = FALSE,
  print_steps = FALSE, print_threshold = FALSE)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{...}{Sets of variables to analyse (see above).}

\item{alpha}{The alpha used for testing.}

\item{show}{Show results (tibbles / plots).}

\item{force_show}{A list of ids of results that should be shown regardless of
\code{show}.}

\item{print_auc}{Print the AUC onto the ROC plot?}

\item{print_points}{Print all data points onto the ROC plot?}

\item{print_steps}{Print only the steps onto the ROC plot?}

\item{print_threshold}{Mark the given threshold in the ROC plot?}
}
\value{
list
}
\description{
Calculate a crosstable or ROC and test metrics (if a cutoff is given) for the
given variables.

It takes sets of the form c(ID, PREDICTOR, RESPONSE, RESPONSE_POS, CUTOFF).
}
