% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{test_wilcoxon_for_groups}
\alias{test_wilcoxon_for_groups}
\title{Calculate Wilcoxon's test (paired or not) for multiple groups.}
\usage{
test_wilcoxon_for_groups(data, column, group, id, set, paired)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{column}{Which column to use.}

\item{group}{The column that parts the variable into groups.}

\item{id}{A column holding ids to identify the same entity in each group.}

\item{set}{A column indicating the two sets to compare values in groups with.}

\item{paired}{Calculate a paired (signed-rank) test or not (rank-sum).}
}
\value{
list
}
\description{
Calculate Wilcoxon's test (paired or not) between two sets for multiple
groups. The groups are specified by column \code{group}, the sets by column
\code{set}, e.g.:

Suppose you had two sets of plants comprising three plants each. One set will
be placed by the window, the other in the corner. You measure height once a
week for ten weeks. Your data looks like:\preformatted{ ID  set  week  height
}

1   1    1     1       6
2   2    1     1       5
3   3    1     1       4
4   4    2     1       5
5   5    2     1       4
6   6    2     1       6
7   1    1     2       7
8   2    1     2       7
9   3    1     2       6
10   4    2     2       6
11   5    2     2       4
12   6    2     2       6
13   1    1     3       8
14   2    1     3       9
15   3    1     3       8
...

You would specify: \code{id = ID}, \code{set = set},  \code{group = week},
\code{column = height} to compare height between both sets (\code{set``) for each week (}group`).
}
