% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{update_homegrown_packages}
\alias{update_homegrown_packages}
\title{Update \code{bioset} and \code{eenv} from github.}
\usage{
update_homegrown_packages()
}
\description{
Update \code{bioset} and \code{eenv} from github.
}
