% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eenv-deprecated.R, R/explore.R
\name{eenv-deprecated}
\alias{eenv-deprecated}
\alias{scan_crosstables}
\alias{calc_crosstables}
\alias{scan_rocs}
\title{Deprecated functions in package \code{een}.}
\usage{
scan_crosstables(data, variables_x, variables_y)

calc_crosstables(data, ...)

scan_rocs(data, predictors, responses, response_pos = TRUE)
}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("-deprecated")}.
}
\section{\code{scan_crosstables}}{

For \code{scan_crosstables} use \code{scan_simple_predictions}.
}

\section{\code{calc_crosstables}}{

For \code{calc_crosstables} use \code{test_simple_predictions}.
}

\section{\code{scan_rocs}}{

For \code{scan_rocs} use \code{test_simple_predictions}.
}

\keyword{internal}
