% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_roc_empiric}
\alias{test_roc_empiric}
\title{Calculate an empiric ROC.}
\usage{
test_roc_empiric(predictors, responses, print_auc = TRUE)
}
\arguments{
\item{predictors}{A vector holding the predictors.}

\item{responses}{A vector holding the responses.}

\item{print_auc}{Print the AUC when creating the plot?}
}
\value{
list
}
\description{
Calculate an empiric ROC.
}
