% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_roc_empiric}
\alias{test_roc_empiric}
\title{Calculate an empiric ROC.}
\usage{
test_roc_empiric(data, predictor, response, print_auc = TRUE)
}
\arguments{
\item{data}{The tibble containing the data.}

\item{predictor}{The column holding the predictors.}

\item{response}{The column holding the responses.}

\item{print_auc}{Print the AUC when creating the plot?}
}
\value{
list
}
\description{
Calculate an empiric ROC.
}
